# hotel-booking-api/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Hotel Booking API
  version: 1.0.0
  description: API de gestão de reservas de hotel (Pedro Cavudissa)
servers:
  - url: http://localhost:9090
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /auth/reset-password/confirm:
    get:
      tags: [Authentication]
      summary: Página reset password
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        200:
          description: HTML page
    post:
      tags: [Authentication]
      summary: Confirmar reset password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string }
      responses:
        200:
          description: Password redefinida
...
  /public/tickets:
    post:
      tags: [Tickets]
      summary: Abrir novo ticket de suporte (Cliente ou Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, message]
              properties:
                subject:
                  type: string
                  example: "Solicitação de alteração de datas"
                message:
                  type: string
                  example: "Gostaria de adiar o meu check-in por motivos de força maior."
                type:
                  type: string
                  enum: [COMPLAINT, ALTERATION, MAINTENANCE]
                  default: COMPLAINT
                  example: "ALTERATION"
                reservationId:
                  type: string
                  format: uuid
                  example: "a6b12345-e89b-12d3-a456-426614174000"
                requestedCheckIn:
                  type: string
                  format: date-time
                  example: "2026-06-15T14:00:00Z"
                requestedCheckOut:
                  type: string
                  format: date-time
                  example: "2026-06-20T12:00:00Z"
                requestedRoomId:
                  type: string
                  format: uuid
                  example: "b7c98765-e89b-12d3-a456-426614174001"
    responses:
      201:
        description: Ticket criado com sucesso.
      400:
        description: Dados inválidos ou sem permissão para a reserva associada.

    get:
      tags: [Tickets]
      summary: Listar os meus tickets (Visão do Cliente)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
          description: Filtrar por estado do ticket
        - in: query
          name: type
          schema:
            type: string
            enum: [COMPLAINT, ALTERATION, MAINTENANCE]
          description: Filtrar por tipo de ticket
      responses:
        200:
          description: Lista de tickets do utilizador retornada com sucesso.
  /public/tickets/{id}:
    get:
      tags: [Tickets]
      summary: Obter detalhes de um ticket específico
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do ticket
      responses:
        200:
          description: Dados detalhados do ticket retornados com sucesso.
        400:
          description: Ticket não encontrado ou acesso negado.
  /public/tickets/{id}/mensagens:
    get:
      tags: [Tickets]
      summary: Listar histórico de mensagens (chat) de um ticket
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do ticket
      responses:
        200:
          description: Histórico de mensagens retornado em ordem cronológica.

    post:
      tags: [Tickets]
      summary: Enviar uma nova mensagem/resposta dentro do ticket
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "Olá, anexei o documento correto. Aguardo a vossa validação."
      responses:
        200:
          description: Mensagem adicionada com sucesso.

  # ─────────────────────────────────────────────────────────────
  # ROTAS ADMINISTRATIVAS DE TICKETS
  # ─────────────────────────────────────────────────────────────
  /public/tickets/admin:
    get:
      tags: [Tickets - Admin]
      summary: Listar todos os tickets do hotel (Backoffice)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        200:
          description: Lista global de tickets gerada com sucesso.
  /public/tickets/admin/{id}:
    patch:
      tags: [Tickets - Admin]
      summary: Atualizar status, responder ou aprovar alterações de reserva
      description: Se o status mudar para 'RESOLVED' em um ticket de 'ALTERATION', a reserva atualiza automaticamente.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
                  example: "RESOLVED"
                response:
                  type: string
                  example: "A sua alteração de estadia foi confirmada e aplicada no nosso sistema."
                assignedToId:
                  type: string
                  description: ID do funcionário responsável pelo atendimento.
      responses:
        200:
          description: Ticket updated com sucesso.
  /public/tickets/admin/stats:
    get:
      tags: [Tickets - Admin]
      summary: Obter métricas quantitativas de tickets por status
      security:
        - bearerAuth: []
      responses:
        200:
          description: Contagem agrupada retornada com sucesso para monitorização do Dashboard.